# docs/openapi.json
{
  "components": {
    "schemas": {
      "AllocationOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "preference_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference Position"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sextet_id": {
            "format": "uuid",
            "title": "Sextet Id",
            "type": "string"
          },
          "sextet_name": {
            "title": "Sextet Name",
            "type": "string"
          },
          "staff_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff Id"
          },
          "staff_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff Name"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trace": {
            "$ref": "#/components/schemas/AllocationTraceOut"
          }
        },
        "required": [
          "id",
          "sextet_id",
          "sextet_name",
          "staff_name",
          "staff_id",
          "run_id",
          "status",
          "kind",
          "preference_position",
          "trace"
        ],
        "title": "AllocationOut",
        "type": "object"
      },
      "AllocationRunDetailOut": {
        "properties": {
          "algorithm_version": {
            "title": "Algorithm Version",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Fingerprint"
          },
          "snapshot": {
            "additionalProperties": true,
            "title": "Snapshot",
            "type": "object"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "algorithm_version",
          "input_fingerprint",
          "snapshot",
          "started_at",
          "finished_at"
        ],
        "title": "AllocationRunDetailOut",
        "type": "object"
      },
      "AllocationRunSummaryOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Fingerprint"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "input_fingerprint"
        ],
        "title": "AllocationRunSummaryOut",
        "type": "object"
      },
      "AllocationTraceOut": {
        "properties": {
          "chosen": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen"
          },
          "fallback_order": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Fallback Order",
            "type": "array"
          },
          "priority_sequence": {
            "title": "Priority Sequence",
            "type": "integer"
          },
          "processing_order": {
            "title": "Processing Order",
            "type": "integer"
          },
          "ranking": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Ranking",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "registration_completed_at": {
            "format": "date-time",
            "title": "Registration Completed At",
            "type": "string"
          },
          "unavailable": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/UnavailableStaffOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable"
          }
        },
        "required": [
          "processing_order",
          "registration_completed_at",
          "priority_sequence",
          "ranking",
          "fallback_order",
          "chosen",
          "reason"
        ],
        "title": "AllocationTraceOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "ResultsOut": {
        "properties": {
          "allocations": {
            "items": {
              "$ref": "#/components/schemas/AllocationOut"
            },
            "title": "Allocations",
            "type": "array"
          },
          "published": {
            "title": "Published",
            "type": "boolean"
          }
        },
        "required": [
          "published",
          "allocations"
        ],
        "title": "ResultsOut",
        "type": "object"
      },
...
      "UnavailableStaffOut": {
        "properties": {
          "sextet_id": {
            "format": "uuid",
            "title": "Sextet Id",
            "type": "string"
          },
          "staff_id": {
            "format": "uuid",
            "title": "Staff Id",
            "type": "string"
          }
        },
        "required": [
          "staff_id",
          "sextet_id"
        ],
        "title": "UnavailableStaffOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "IF-Arbitra",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/rounds/{round_id}/allocate": {
      "post": {
        "operationId": "process_api_admin_rounds__round_id__allocate_post",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Round Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationRunSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Process",
        "tags": [
          "Alocação"
        ]
      }
    },
...
    "/api/admin/runs/{run_id}": {
      "get": {
        "operationId": "run_detail_api_admin_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationRunDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Run Detail",
        "tags": [
          "Auditoria"
        ]
      }
    },
...
    "/api/rounds/{round_id}/results": {
      "get": {
        "operationId": "results_api_rounds__round_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Round Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Results",
        "tags": [
          "Alocação"
        ]
      }
    },
...
  }
}